add active column to User model; enforce user ownership in task and user deletion; update task and user services to utilize token payload

+++ src/tasks/tasks.service.ts
@@ -3,6 +3,7 @@ import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { PaginationDto } from 'src/common/dto/pagination.dto';
+import { PayloadTokenDto } from 'src/auth/dto/payload-token.dto';
 
 @Injectable()
 export class TasksService {
@@ -39,13 +40,14 @@ export class TasksService {
     }
   }
 
-  async create(createTaskDto: CreateTaskDto) {
+  async create(createTaskDto: CreateTaskDto, tokenPayload: PayloadTokenDto) {
     try {
       const newTask = await this.prisma.task.create({
         data: {
           name: createTaskDto.name,
           description: createTaskDto.description,
           completed: false,
+          userId: tokenPayload.sub,
         },
       });
       return newTask;
@@ -55,13 +57,16 @@ export class TasksService {
     }
   }
 
-  async update(id: number, updateTaskDto: UpdateTaskDto) {
+  async update(id: number, updateTaskDto: UpdateTaskDto, tokenPayload: PayloadTokenDto) {
     try {
       const findTask = await this.prisma.task.findFirst({
         where: { id },
       });
       if (!findTask) {
         throw new HttpException(`Task with id ${id} not found`, HttpStatus.NOT_FOUND);
+      }
+      if (findTask.userId !== tokenPayload.sub) {
+        throw new HttpException('You can only update your own tasks', HttpStatus.FORBIDDEN);
       }
       const updatedTask = await this.prisma.task.update({
         where: { id },
@@ -78,13 +83,16 @@ export class TasksService {
     }
   }
 
-  async delete(id: number) {
+  async delete(id: number, tokenPayload: PayloadTokenDto) {
     try {
       const findTask = await this.prisma.task.findFirst({
         where: { id },
       });
       if (!findTask) {
         throw new HttpException(`Task with id ${id} not found`, HttpStatus.NOT_FOUND);
+      }
+      if (findTask.userId !== tokenPayload.sub) {
+        throw new HttpException('You can only delete your own tasks', HttpStatus.FORBIDDEN);
       }
       const deletedTask = await this.prisma.task.delete({
         where: { id },
